refactor: atualização de exception

+++ backend-orange/src/main/java/com/hackathon/backendorange/service/ProjectService.java
@@ -3,6 +3,8 @@ package com.hackathon.backendorange.service;
 import com.cloudinary.Cloudinary;
 import com.hackathon.backendorange.controller.assembler.ProjectAssembler;
 import com.hackathon.backendorange.dto.ProjectDTO;
+import com.hackathon.backendorange.exception.ProjectIdNotFoundException;
+import com.hackathon.backendorange.exception.UserNotFoundException;
 import com.hackathon.backendorange.model.Project;
 import com.hackathon.backendorange.model.User;
 import com.hackathon.backendorange.repository.ProjectRepository;
@@ -12,10 +14,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.List;
-import java.util.Map;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.*;
 
 @Service
 public class ProjectService {
@@ -43,15 +42,16 @@ public class ProjectService {
             User userExists = user.get();
             return repository.getUserProjects(userExists.getId());
         } else {
-            throw new RuntimeException("Usuário não encontrado!");
+            throw new UserNotFoundException();
         }
     }
 
     public ProjectDTO saveProject(ProjectDTO projectDTO, MultipartFile file) throws IOException {
 
         if (projectDTO != null && projectDTO.getUser().getId() != null) {
-            User user = userRepository.findById(projectDTO.getUser().getId()).orElseThrow(() -> new RuntimeException(
-                    "Usuário não encontrado!"));
+            User user =
+                    userRepository.findById(projectDTO.getUser().getId()).orElseThrow(() ->
+                            new UserNotFoundException());
 
             Map imageInfo = uploadImage(file);
             String imageUrl = imageInfo.get("url").toString();
@@ -60,6 +60,7 @@ public class ProjectService {
             Project projectEntity = projectAssembler.toEntity(projectDTO);
             projectEntity.setImage(imageUrl);
             projectEntity.setImage_id(imageId);
+            projectEntity.setImage_originalName(file.getOriginalFilename());
 
             Project savedProject = repository.save(projectEntity);
             ProjectDTO savedProjectDTO = projectAssembler.toDTO(savedProject);
@@ -73,10 +74,9 @@ public class ProjectService {
 
     public Optional<ProjectDTO> updateProject(Long id, ProjectDTO projectDTO, MultipartFile file) throws IOException {
         if (projectDTO != null && projectDTO.getUser().getId() != null && id != null) {
-            User user = userRepository.findById(projectDTO.getUser().getId()).orElseThrow(() -> new RuntimeException(
-                    "Usuário não encontrado!"));
-            Project projectExists = repository.findById(id).orElseThrow(() -> new RuntimeException(
-                    "Projeto não encontrado!"));
+            User user = userRepository.findById(projectDTO.getUser().getId()).orElseThrow(() ->
+                    new UserNotFoundException());
+            Project projectExists = repository.findById(id).orElseThrow(() -> new ProjectIdNotFoundException());
 
             projectExists.setTitulo(projectDTO.getTitulo());
             projectExists.setDescricao(projectDTO.getDescricao());
@@ -84,17 +84,20 @@ public class ProjectService {
             projectExists.setLinks(projectDTO.getLinks());
             projectExists.setUser(projectDTO.getUser());
 
-            //delete -> imagem antiga
-            String imageIdExists = projectDTO.getImage_id();
-            deleteImage(imageIdExists);
+            if (!Objects.equals(projectExists.getImage_originalName(), file.getOriginalFilename())){
+                //delete -> imagem antiga
+                String imageIdExists = projectDTO.getImage_id();
+                deleteImage(imageIdExists);
 
-            //upload -> imagem nova
-            Map imageInfo = uploadImage(file);
-            String imageUrl = imageInfo.get("url").toString();
-            String image_id = imageInfo.get("public_id").toString();
+                //upload -> imagem nova
+                Map imageInfo = uploadImage(file);
+                String imageUrl = imageInfo.get("url").toString();
+                String image_id = imageInfo.get("public_id").toString();
 
-            projectExists.setImage(imageUrl);
-            projectExists.setImage_id(image_id);
+                projectExists.setImage(imageUrl);
+                projectExists.setImage_id(image_id);
+            }
+
 
             Project savedProject = repository.save(projectExists);
             ProjectDTO savedProjectDTO = projectAssembler.toDTO(savedProject);
@@ -107,8 +110,7 @@ public class ProjectService {
 
     public Project deleteProject(Long id) throws IOException {
         if (id != null) {
-            Project projectExists = repository.findById(id).orElseThrow(() -> new RuntimeException(
-                    "Projeto não encontrado!"));
+            Project projectExists = repository.findById(id).orElseThrow(() -> new ProjectIdNotFoundException());
 
             String image_id = projectExists.getImage_id();
             deleteImage(image_id);
